be sure that the ... arguments to wh_request are correctly casted to uint64_t

+++ tests/orwl_wait_list.c
@@ -92,7 +92,7 @@ DEFINE_THREAD(arg_t) {
     /* the postion where we put the callback and that we acquire */
     size_t pacq = orwl_mynum + (orwl_phase % 2)*orwl_np;
     orwl_state ostate =
-      orwl_wq_request(&location, handle + preq, 1);
+      orwl_wq_request(&location, handle + preq, (uint64_t)1);
     report(!orwl_mynum,  "req, handle %zu, %s",
            preq, orwl_state_getname(ostate));
     /**/
